Splits out Slack message formatter into separate class behind ISlackFormatter.
