+++ src/Narochno.Serilog.Slack/Formatting/ISlackFormatter.cs
@@ -1,0 +1,11 @@
+﻿using Narochno.Slack.Entities;
+using Serilog.Events;
+using System.Collections.Generic;
+
+namespace Narochno.Serilog.Slack.Formatting
+{
+    public interface ISlackFormatter
+    {
+        Message CreateMessage(IEnumerable<LogEvent> logEvents);
+    }
+}
