+++ src/Narochno.Serilog.Slack/project.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.0.4",
+  "version": "1.1.0",
 
   "dependencies": {
     "Narochno.Slack": "1.4.2",
